Java: Add default values for builders (#363)
* Get builder's defaults and fix safeguards

* Set default values

* Update tests

* Add list default case

* Fix enum values

* Update tests

* Fix defaults for anonymous structs

+++ testdata/jennies/rawtypes/struct_with_scalar_fields/JavaRawTypes/basic/SomeStruct.java
@@ -15,11 +15,11 @@ public class SomeStruct {
     public String fieldStringWithConstantValue;
     public Float fieldFloat32;
     public Double fieldFloat64;
-    public Byte fieldUint8;
+    public Integer fieldUint8;
     public Short fieldUint16;
     public Integer fieldUint32;
     public Long fieldUint64;
-    public Byte fieldInt8;
+    public Integer fieldInt8;
     public Short fieldInt16;
     public Integer fieldInt32;
     public Long fieldInt64;
